feat: add support for both postgres and sqlite

+++ plugins/qeta-backend/migrations/20230314-search-index.js
@@ -2,14 +2,18 @@
  * @param {import('knex').Knex} knex
  */
 exports.up = async function up(knex) {
-  await knex.schema.raw(
-    `CREATE INDEX questions_search_content_index ON questions USING GIN (to_tsvector('english', title || ' ' || content))`,
-  );
+  if (knex.client.config.client === 'pg') {
+    await knex.schema.raw(
+      `CREATE INDEX questions_search_content_index ON questions USING GIN (to_tsvector('english', title || ' ' || content))`,
+    );
+  }
 };
 
 /**
  * @param {import('knex').Knex} knex
  */
 exports.down = async function down(knex) {
-  await knex.schema.raw(`DROP INDEX questions_search_content_index`);
+  if (knex.client.config.client === 'pg') {
+    await knex.schema.raw(`DROP INDEX questions_search_content_index`);
+  }
 };
